move VectorDistance related traits into vectors
