+++ et/src/spann/bulk_load.rs
@@ -3,7 +3,6 @@ use std::{fs::File, io, num::NonZero, path::PathBuf, sync::Arc};
 use clap::Args;
 use easy_tiger::{
     bulk::{self, BulkLoadBuilder},
-    distance::VectorSimilarity,
     graph::{GraphConfig, GraphLayout, GraphSearchParams},
     input::{DerefVectorStore, SubsetViewVectorStore, VectorStore},
     kmeans::{iterative_balanced_kmeans, Params},
@@ -13,7 +12,7 @@ use easy_tiger::{
         },
         IndexConfig, TableIndex,
     },
-    vectors::F32VectorCoding,
+    vectors::{F32VectorCoding, VectorSimilarity},
 };
 use histogram::Histogram;
 use rand_xoshiro::{rand_core::SeedableRng, Xoshiro128PlusPlus};
