+++ src/lib.rs
@@ -12,7 +12,6 @@ pub mod graph;
 pub mod graph_clustering;
 pub mod input;
 pub mod kmeans;
-pub mod query_distance;
 pub mod search;
 pub mod spann;
 pub mod vectors;
