+++ src/vectors/i8naive.rs
@@ -1,4 +1,4 @@
-use crate::vectors::F32VectorCoder;
+use crate::vectors::{F32VectorCoder, VectorDistance, VectorSimilarity};
 
 #[derive(Debug, Copy, Clone)]
 pub struct I8NaiveVectorCoder;
@@ -21,5 +21,40 @@ impl F32VectorCoder for I8NaiveVectorCoder {
 
     fn byte_len(&self, dimensions: usize) -> usize {
         dimensions + std::mem::size_of::<f32>()
+    }
+}
+
+#[derive(Debug, Copy, Clone)]
+pub struct I8NaiveDistance(pub(crate) VectorSimilarity);
+
+impl I8NaiveDistance {
+    fn unpack(raw: &[u8]) -> (&[i8], f32) {
+        assert!(raw.len() >= std::mem::size_of::<f32>());
+        let (vector_bytes, norm_bytes) = raw.split_at(raw.len() - std::mem::size_of::<f32>());
+        (
+            bytemuck::cast_slice(vector_bytes),
+            f32::from_le_bytes(norm_bytes.try_into().unwrap()),
+        )
+    }
+}
+
+impl VectorDistance for I8NaiveDistance {
+    fn distance(&self, query: &[u8], doc: &[u8]) -> f64 {
+        let (qv, qnorm) = Self::unpack(query);
+        let (dv, dnorm) = Self::unpack(doc);
+        let divisor = i8::MAX as f32 * i8::MAX as f32;
+        // XXX this might not be correct, IIRC I need to scale dot based on the norms like in the
+        // scaled uniform implementation.
+        // NB: we may be able to accelerate this further with manual SIMD implementations.
+        let dot = qv
+            .iter()
+            .zip(dv.iter())
+            .map(|(q, d)| *q as i32 * *d as i32)
+            .sum::<i32>() as f64
+            / divisor as f64;
+        match self.0 {
+            VectorSimilarity::Dot => (-dot + 1.0) / 2.0,
+            VectorSimilarity::Euclidean => qnorm as f64 + dnorm as f64 - (2.0 * dot),
+        }
     }
 }
